Style links to the current page differently
With jQuery magic, add a class to all links in the main content and the
sidebar and style it to look different from other links (in particular,
to have black color and bold font).  This is aimed to improve navigation
with links in the sidebar as you can clearly see what page you are
currently on.

Also add that feature to the CHANGELOG.